Use first 16 bytes for IV
The protocol seems to use the first 16 bytes of the previous message
as the IV for the next message. This was previously using the last 16
bytes of the previous message as the IV. 

Discovered as the second 16 bytes of a message wound up decrypting fine,
only the first 16 were borked which pointed to a broken IV, but it was
being passed fine, just generated incorrectly.