protocol: Respond with the current time
Not sure if this is set correctly on the Photon currently, still need
to verify. Probably with a periodic function on the Photon to send
the current time.

I guess this kind of means I should make a dev program for the photon
as this point that exercises a bunch of the cloud functions!